reorganize: put interval type with other datetimes
